move validateStartDate function to helpers
