+++ src/adapters/leumi/scrape-transactions.js
@@ -7,7 +7,7 @@ import {
 import { SHEKEL_CURRENCY, NORMAL_TXN_TYPE, TRANSACTION_STATUS } from '../../constants';
 import { mapAccounts, navigateToAccountTransactions } from './adapterHelpers/accounts';
 import { DATE_FORMAT } from './definitions';
-import validateStartDate from './adapterHelpers/scraping';
+import { validateStartDate } from '../../helpers/dates';
 
 const NO_TRANSACTION_IN_DATE_RANGE_TEXT = 'לא קיימות תנועות מתאימות על פי הסינון שהוגדר';
